feat(fork-choice): add applyScoreChanges, findHead, and ePBS viability logic
Implement weight propagation (two-pass backward iteration), head
selection via best_descendant chains, and Gloas-specific node
comparison with payload status tiebreakers.

Key additions to ProtoArray:
- applyScoreChanges: proposer boost undo/redo, checked arithmetic,
  weight back-propagation, best child/descendant recomputation
- findHead: O(1) head via justified -> best_descendant chain
- nodeIsViableForHead: FFG filter with pull-up justification
- isFinalizedRootOrDescendant: 4 fast-path checkpoint checks +
  parent chain walk fallback
- getAncestor: slot-based ancestor lookup via parent_root chain
- shouldExtendPayload / isPayloadTimely: PTC vote counting for
  Gloas payload timeliness
- getPayloadStatusTiebreaker: EMPTY vs FULL comparison for
  same-root Gloas nodes
- propagateValidExecutionStatusByIndex: upward EL status promotion

Also changes ProtoNode index fields from ?usize to ?u32 for
TigerStyle explicit-size compliance.

+++ src/fork_choice/proto_node.zig
@@ -251,18 +251,18 @@ pub const ProtoNode = struct {
     // ── DAG metadata ──
 
     /// Index of parent node in the nodes array. null for the root.
-    parent: ?usize = null,
+    parent: ?u32 = null,
 
     /// LMD-GHOST weight: sum of effective balances of validators
     /// whose latest vote is for this subtree.
     weight: i64 = 0,
 
     /// Index of the highest-weight child.
-    best_child: ?usize = null,
+    best_child: ?u32 = null,
 
     /// Index of the best leaf reachable from this node.
     /// findHead: justified_root -> bestDescendant in O(1).
-    best_descendant: ?usize = null,
+    best_descendant: ?u32 = null,
 
     /// Create a ProtoNode from a ProtoBlock, copying all matching fields.
     pub fn fromBlock(block: ProtoBlock) ProtoNode {
